+++ lib/processor/condition/static.go
@@ -41,6 +41,10 @@ Static is a condition that always resolves to the same static boolean value.`,
 // Static is a condition that always returns a static boolean value.
 type Static struct {
 	value bool
+
+	mCount metrics.StatCounter
+	mTrue  metrics.StatCounter
+	mFalse metrics.StatCounter
 }
 
 // NewStatic returns a Static condition.
@@ -49,6 +53,10 @@ func NewStatic(
 ) (Type, error) {
 	return &Static{
 		value: conf.Static,
+
+		mCount: stats.GetCounter("count"),
+		mTrue:  stats.GetCounter("true"),
+		mFalse: stats.GetCounter("false"),
 	}, nil
 }
 
@@ -56,6 +64,12 @@ func NewStatic(
 
 // Check attempts to check a message part against a configured condition.
 func (s *Static) Check(msg types.Message) bool {
+	s.mCount.Incr(1)
+	if s.value {
+		s.mTrue.Incr(1)
+	} else {
+		s.mFalse.Incr(1)
+	}
 	return s.value
 }