+++ lib/processor/condition/xor.go
@@ -1,18 +1,18 @@
 // Copyright (c) 2018 Ashley Jeffs
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software or associated documentation files (the "Software"), to deal
+// of this software and associated documentation files (the "Software"), to deal
 // in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, or/or sell
-// copies of the Software, or to permit persons to whom the Software is
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions:
 //
-// The above copyright orice or this permission orice shall be included in
+// The above copyright notice and this permission notice shall be included in
 // all copies or substantial portions of the Software.
 //
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT or LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE or NONINFRINGEMENT. IN NO EVENT SHALL THE
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 // AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
@@ -65,6 +65,10 @@ func NewXorConfig() XorConfig {
 // Xor is a condition that returns the logical xor of all children.
 type Xor struct {
 	children []Type
+
+	mCount metrics.StatCounter
+	mTrue  metrics.StatCounter
+	mFalse metrics.StatCounter
 }
 
 // NewXor returns an Xor condition.
@@ -73,7 +77,7 @@ func NewXor(
 ) (Type, error) {
 	children := []Type{}
 	for i, childConf := range conf.Xor {
-		ns := fmt.Sprintf("xor.%v", i)
+		ns := fmt.Sprintf("%v", i)
 		child, err := New(childConf, mgr, log.NewModule("."+ns), metrics.Namespaced(stats, ns))
 		if err != nil {
 			return nil, fmt.Errorf("failed to create child '%v': %v", childConf.Type, err)
@@ -82,6 +86,10 @@ func NewXor(
 	}
 	return &Xor{
 		children: children,
+
+		mCount: stats.GetCounter("count"),
+		mTrue:  stats.GetCounter("true"),
+		mFalse: stats.GetCounter("false"),
 	}, nil
 }
 
@@ -89,14 +97,21 @@ func NewXor(
 
 // Check attempts to check a message part against a configured condition.
 func (c *Xor) Check(msg types.Message) bool {
+	c.mCount.Incr(1)
 	hadTrue := false
 	for _, child := range c.children {
 		if child.Check(msg) {
 			if hadTrue {
+				c.mFalse.Incr(1)
 				return false
 			}
 			hadTrue = true
 		}
+	}
+	if hadTrue {
+		c.mTrue.Incr(1)
+	} else {
+		c.mFalse.Incr(1)
 	}
 	return hadTrue
 }