+++ lib/util/http/client/type_test.go
@@ -123,6 +123,30 @@ func TestHTTPClientSendBasic(t *testing.T) {
 	}
 }
 
+func TestHTTPClientDropOn(t *testing.T) {
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte(`{"foo":"bar"}`))
+	}))
+	defer ts.Close()
+
+	conf := NewConfig()
+	conf.URL = ts.URL + "/testpost"
+	conf.DropOn = []int{400}
+
+	h, err := New(conf)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	testMsg := message.New([][]byte{[]byte(`{"bar":"baz"}`)})
+
+	resMsg, err := h.Send(testMsg)
+	if err == nil {
+		t.Errorf("Expected error, received: %s", message.GetAllBytes(resMsg))
+	}
+}
+
 func TestHTTPClientSendInterpolate(t *testing.T) {
 	nTestLoops := 1000