+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - API: The `types.Processor` interface has been changed in order to add lifetime
   cleanup methods (added `CloseAsync` and `WaitForClose`). For the overwhelming
   majority of processors these functions will be no-ops.
+- More consistent `condition` metrics.
 
 ## 0.39.2 - 2018-12-07